Fix call of the 2+ sfill operations
Call of the second and next sfill job was crashing because of a sort
of "deadlock": a GsdAsyncOperation is locked when running, and this
lock prevents any reuse of the operation. Unfortunately, we was
reusing it without any special care just before it unlocks itself.
Now we trigger the call in a timeout function when the previous
operation is ready.