Fix support for Automake 1.12
Automake 1.12 now overrides warnings implied by the chosen strictness
even if that strictness appears after the warning flag.  In practice,
specifying "foreign" after "-Wall" no longer disables portability
warnings, but we need to disable them for gnome-doc-utils -- and not to
need AM_PROG_AR when using libtool, which is required since 1.11.2
when the -Wextra-portability flag is enabled.

So, explicitly disable portability warnings for fix support for
Automake 1.12.