Fix signal re-emission
g_signal_emit_by_name() do not take a separate detail argument, we only
have been lucky it worked for progress report.  This was only because
the x86-64 calling convention passes doubles in different registers
than integers and pointers, so as the erroneous extra argument was an
integer it didn't conflict on x86-64.

However, the one for the "finished" signal was just broken, as the
arguments (no doubles) were offset by one (on x86-64, probably worse on
stack-based architectures), leading to the "success" parameter being
interpreted as the "message" one.
It might however not have been too problematic in practice, because the
erroneous extra leading argument was `0`, which also was the only
possible value of the `success` variable used as second argument.  As
0 is a proper NULL pointer, the call may have been equivalent to
`(..., FALSE, NULL)`, which, although not the expected arguments here
(second one should likely not be `NULL`), is a valid argument list for
the signal (`message` argument is allowed to be `NULL`).
This might even also apply to stack-based calling convention if the
sizes and offset of integers and pointer are compatible, which is not
unlikely.

Anyway, remove the erroneous extra leading argument fixing all this
craziness.