Pause the operation while the cancel confirmation dialog is shown
This still allows the confirmation dialog yet makes it effective faster
as the operation will be paused waiting for an answer, effectively
canceling it at the time the button was initially pressed if confirmed.

This also follows better the wording of that dialog that suggests the
operation is paused by suggesting one can "resume it" by not canceling
it.