Fix resolving path to virtual (desktop shortcut) locations
Do not consider every URI with an x-nautilus-desktop scheme as pointing
directly to the desktop, as "Computer", "Trash" and other virtual
shortcuts share this scheme on the desktop.

This fixes handling of all virtual shortcuts pointing to simple local
locations, like drives and home directory.  It doesn't handle trash
yet, because it's a lot less straightforward and Nautilus doesn't help
much on that.

Fixes https://labs.riseup.net/code/issues/9710